Glossary-Scraper/requirements.txt

requests==2.28.1
beautifulsoup4==4.12.2
python-docx==0.8.11
html5lib==1.1
